refactor: separar renderizado de catalogo de paginacion en UI
El archivo catalogDOMManip contaba con funciones referidas tanto a
renderizado del catalogo (tarjetas producto, poblar el catalogo) como
con funciones de paginacion (crear botones de pagina con eventos).
Fueron separados para facilitar su mantenimiento.
