+++ views/catalogo.hbs
@@ -1,4 +1,5 @@
-  <script src="/static/javascripts/catalogDOMManip.js" type="module"></script>
+  <script src="/static/javascripts/productsUI.js" type="module"></script>
+  <script src="/static/javascripts/pagination.js" type="module"></script>
   <script src="/static/javascripts/populateCatalog.js" type="module"></script>
   <script src="/static/javascripts/createProduct.js" type="module"></script>
